requirements/extra.txt
docutils
-e hg+https://bitbucket.org/brodie/django-piston#egg=django-piston
-e git://github.com/openpolis/piston_mini_client.git#egg=piston_mini_client
django-debug-toolbar
fabric
lorem-ipsum-generator
-e git://github.com/openpolis/rst2pdf-patched-docutils-0.8#egg=rst2pdf
